requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pillow==10.1.0
pyheif==0.7.1
python-multipart==0.0.6
exifread==3.0.0
watchdog==3.0.0
pydantic==2.5.2
python-dotenv==1.0.0
aiofiles==23.2.1 